Add environment configuration for MongoDB and Kafka, update application properties

+++ travel-agency-query-side/src/main/java/com/rzodeczko/application/command/UpdateAvailabilityCommand.java
@@ -2,7 +2,7 @@ package com.rzodeczko.application.command;
 
 import java.time.LocalDate;
 
-public record UpdateAvailabilityCommand(long hotelId, LocalDate date, int occupied) {
+public record UpdateAvailabilityCommand(long hotelId, LocalDate date, long occupied) {
     public UpdateAvailabilityCommand {
         if (date == null) {
             throw new IllegalArgumentException("Date is required");
